UC-165 Model design V4 - Adjusting counter helper
