--- counter/byMonth.js
@@ -1,254 +1,0 @@
-const { byMonth, Request } = require('./getter')
-const { table } = require('table')
-const { LocalDate } = require('@js-joda/core')
-
-/**
- * ╔════════════════════════════════════════════════════════════════╗
- * ║ <<Aggregator step I>>                                          ║
- * ║                                                                ║
- * ║ Host=http://localhost:8081                                     ║
- * ║ IdentityId=12539f76-c7e9-47d6-b37b-4b59380721ac                ║
- * ║ Category=verification                                          ║
- * ║ Subcategory=entry-a                                            ║
- * ║ Month=FEBRUARY                                                 ║
- * ╟────────────┬────────────┬────────────┬────────────┬────────────╢
- * ║ Year       │ Month      │ Day        │ Hour       │ Count      ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 0          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 1          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 2          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 3          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 4          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 5          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 6          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 7          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 8          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 9          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 10         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 11         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 12         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 13         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 14         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 15         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 16         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 17         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 18         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 19         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 20         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 21         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 22         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 1          │ 23         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 0          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 1          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 2          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 3          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 4          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 5          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 6          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 7          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 8          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 9          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 10         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 11         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 12         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 13         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 14         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 15         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 16         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 17         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 18         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 19         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 20         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 21         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 22         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 2          │ 23         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 0          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 1          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 2          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 3          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 4          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 5          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 6          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 7          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 8          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 9          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 10         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 11         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 12         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 13         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 14         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 15         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 16         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 17         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 18         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 19         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 20         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 21         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 22         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 3          │ 23         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 0          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 1          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 2          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 3          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 4          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 5          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 6          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 7          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 8          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 9          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 10         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 11         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 12         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 13         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 14         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 15         │ 1000000    ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 16         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 17         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 18         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 19         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 20         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 21         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 22         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 23         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ Year       │ Month      │ Day        │ Hour       │ Count      ║
- * ╚════════════╧════════════╧════════════╧════════════╧════════════╝
- */
-
-const config = require('./config.json')
-const request = new Request(
-  config.host,
-  config.token,
-  '12539f76-c7e9-47d6-b37b-4b59380721ac',
-  'verification',
-  '2022-02-04',
-  '-2',
-  'entry-a'
-)
-
-byMonth(request)
-  .then(function (response) {
-    const data = [['Year', 'Month', 'Day', 'Hour', 'Count']]
-    for (const i of response) {
-      data.push([i.year, i.month, i.day, i.hour, i.count])
-    }
-    data.push(['Year', 'Month', 'Day', 'Hour', 'Count'])
-    const config = {
-      columnDefault: {
-        width: 10
-      },
-      header: {
-        alignment: 'left',
-        content: '<<Aggregator step I>>' +
-                    '\n' +
-                    '\nHost=' + request.host +
-                    '\nIdentityId=' + request.identityId +
-                    '\nCategory=' + request.queryParams.cat +
-                    '\nSubcategory=' + request.queryParams.sub_cat +
-                    '\nMonth=' + LocalDate.parse(request.queryParams.date).atStartOfDay().month().name()
-      }
-    }
-
-    console.log(table(data, config))
-  })
-  .catch(function (error) {
-    console.log(error)
-  })
-  .then(function () {
-    // always executed
-  })
