--- counter/byMonthAndDay.js
@@ -1,115 +1,0 @@
-const { Request, byDayAndHour } = require('./getter')
-const { table } = require('table')
-const { LocalDate } = require('@js-joda/core')
-
-/**
- * ╔════════════════════════════════════════════════════════════════╗
- * ║ <<Aggregator step I>>                                          ║
- * ║                                                                ║
- * ║ Host=http://localhost:8081                                     ║
- * ║ IdentityId=12539f76-c7e9-47d6-b37b-4b59380721ac                ║
- * ║ Category=verification                                          ║
- * ║ Subcategory=entry-an                                           ║
- * ║ Month=FEBRUARY                                                 ║
- * ║ Day=4                                                          ║
- * ╟────────────┬────────────┬────────────┬────────────┬────────────╢
- * ║ Year       │ Month      │ Day        │ Hour       │ Count      ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 0          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 1          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 2          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 3          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 4          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 5          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 6          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 7          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 8          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 9          │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 10         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 11         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 12         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 13         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 14         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 15         │ 1000000    ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 16         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 17         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 18         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 19         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 20         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 21         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 22         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ 2022       │ 2          │ 4          │ 23         │ 0          ║
- * ╟────────────┼────────────┼────────────┼────────────┼────────────╢
- * ║ Year       │ Month      │ Day        │ Hour       │ Count      ║
- * ╚════════════╧════════════╧════════════╧════════════╧════════════╝
- */
-
-const config = require('./config.json')
-const request = new Request(
-  config.host,
-  config.token,
-  '12539f76-c7e9-47d6-b37b-4b59380721ac',
-  'verification',
-  '2022-02-04',
-  '-2',
-  'entry-a'
-)
-
-byDayAndHour(request)
-  .then(function (response) {
-    const data = [['Year', 'Month', 'Day', 'Hour', 'Count']]
-
-    for (const i of response) {
-      data.push([i.year, i.month, i.day, i.hour, i.count])
-    }
-
-    data.push(['Year', 'Month', 'Day', 'Hour', 'Count'])
-
-    const config = {
-      columnDefault: {
-        width: 10
-      },
-      header: {
-        alignment: 'left',
-        content: '<<Aggregator step I>>' +
-                    '\n' +
-                    '\nHost=' + request.host +
-                    '\nIdentityId=' + request.identityId +
-                    '\nCategory=' + request.queryParams.cat +
-                    '\nSubcategory=' + request.queryParams.sub_cat +
-                    '\nMonth=' + LocalDate.parse(request.queryParams.date).atStartOfDay().month().name() +
-                    '\nDay=' + LocalDate.parse(request.queryParams.date).atStartOfDay().dayOfMonth()
-      }
-    }
-
-    console.log(table(data, config))
-  })
-  .catch(function (error) {
-    console.log(error)
-  })
-  .then(function () {
-    // always executed
-  })
